Comment on FE 595
-Does the script error out and fail if the cURL call fails?
It does not error out and fail if the call to the URL fails, however, the website may throw an error if it is unavailable. 
-Can the script augment an existing file or does it always create a new one?
The script will write over the file with the same name in the directory it is set to. 
-Is the script easy to read/edit/configure?
Yes, the code is separated into readable chunks that make would make editing easy. 
-Will the script run if imported? (It should only run if called, not if imported.)
The __main__ at the end of the code will make sure the code is only run if called
-How many times does the script write to the output file? Can we make it write to the disk just once?
The script only writes to the output file once, it will overwrite the file that is in that directory with that name
-If it does all of these things, am I acknowledging that the script does these things successfully?
Yes.